Remove TODOs and move code to separate files
